recorrect startupscript

+++ src/main/java/com/one211/restapi/model/StartupScript.java
@@ -3,4 +3,4 @@ package com.one211.restapi.model;
 
 import java.time.LocalDateTime;
 
-public record StartupScript(Long orgId, String clusterName, String version, String content, LocalDateTime createdAt, LocalDateTime updatedAt) {}
+public record StartupScript(Integer orgId, String clusterName, String version, String content, LocalDateTime createdAt, LocalDateTime updatedAt) {}
